Use termios via CGO
